effects: move Blood
