+++ src/game/effects/blood.h
@@ -1,0 +1,7 @@
+#pragma once
+
+#include <stdint.h>
+
+int16_t Effect_Blood(
+    int32_t x, int32_t y, int32_t z, int16_t speed, int16_t direction,
+    int16_t room_num);
