+++ src/game/larafire.c
@@ -6,9 +6,9 @@
 #include "config.h"
 #include "game/control.h"
 #include "game/draw.h"
+#include "game/effects/blood.h"
 #include "game/input.h"
 #include "game/inv.h"
-#include "game/objects/effects/blood.h"
 #include "game/objects/effects/ricochet.h"
 #include "game/random.h"
 #include "game/sound.h"
@@ -605,7 +605,7 @@ void HitTarget(ITEM_INFO *item, GAME_VECTOR *hitpos, int32_t damage)
     item->hit_points -= damage;
     item->hit_status = 1;
 
-    Blood_Spawn(
+    Effect_Blood(
         hitpos->x, hitpos->y, hitpos->z, item->speed, item->pos.y_rot,
         item->room_number);
