+++ src/game/objects/ai/alligator.c
@@ -4,10 +4,10 @@
 #include "game/box.h"
 #include "game/collide.h"
 #include "game/control.h"
+#include "game/effects/blood.h"
 #include "game/items.h"
 #include "game/lot.h"
 #include "game/objects/ai/alligator.h"
-#include "game/objects/effects/blood.h"
 #include "global/vars.h"
 
 void SetupAlligator(OBJECT_INFO *obj)
@@ -122,7 +122,7 @@ void AlligatorControl(int16_t item_num)
 
         if (info.bite && item->touch_bits) {
             if (item->required_anim_state == ALLIGATOR_EMPTY) {
-                CreatureEffect(item, &g_CrocodileBite, Blood_Spawn);
+                CreatureEffect(item, &g_CrocodileBite, Effect_Blood);
                 g_LaraItem->hit_points -= ALLIGATOR_BITE_DAMAGE;
                 g_LaraItem->hit_status = 1;
                 item->required_anim_state = ALLIGATOR_SWIM;
