+++ src/game/objects/ai/ape.c
@@ -2,8 +2,8 @@
 
 #include "game/box.h"
 #include "game/collide.h"
+#include "game/effects/blood.h"
 #include "game/lot.h"
-#include "game/objects/effects/blood.h"
 #include "game/random.h"
 #include "global/vars.h"
 
@@ -198,7 +198,7 @@ void ApeControl(int16_t item_num)
 
         case APE_ATTACK1:
             if (!item->required_anim_state && (item->touch_bits & APE_TOUCH)) {
-                CreatureEffect(item, &g_ApeBite, Blood_Spawn);
+                CreatureEffect(item, &g_ApeBite, Effect_Blood);
                 g_LaraItem->hit_points -= APE_ATTACK_DAMAGE;
                 g_LaraItem->hit_status = 1;
                 item->required_anim_state = APE_STOP;
