+++ src/game/objects/ai/bear.c
@@ -2,8 +2,8 @@
 
 #include "game/box.h"
 #include "game/collide.h"
+#include "game/effects/blood.h"
 #include "game/lot.h"
-#include "game/objects/effects/blood.h"
 #include "game/random.h"
 #include "global/types.h"
 #include "global/vars.h"
@@ -184,7 +184,7 @@ void BearControl(int16_t item_num)
 
         case BEAR_ATTACK1:
             if (!item->required_anim_state && (item->touch_bits & BEAR_TOUCH)) {
-                CreatureEffect(item, &g_BearHeadBite, Blood_Spawn);
+                CreatureEffect(item, &g_BearHeadBite, Effect_Blood);
                 g_LaraItem->hit_points -= BEAR_ATTACK_DAMAGE;
                 g_LaraItem->hit_status = 1;
                 item->required_anim_state = BEAR_STOP;
