+++ src/game/objects/ai/lion.c
@@ -2,8 +2,8 @@
 
 #include "game/box.h"
 #include "game/collide.h"
+#include "game/effects/blood.h"
 #include "game/lot.h"
-#include "game/objects/effects/blood.h"
 #include "game/random.h"
 #include "global/vars.h"
 
@@ -169,7 +169,7 @@ void LionControl(int16_t item_num)
         case LION_ATTACK2:
             if (item->required_anim_state == LION_EMPTY
                 && (item->touch_bits & LION_TOUCH)) {
-                CreatureEffect(item, &g_LionBite, Blood_Spawn);
+                CreatureEffect(item, &g_LionBite, Effect_Blood);
                 g_LaraItem->hit_points -= LION_BITE_DAMAGE;
                 g_LaraItem->hit_status = 1;
                 item->required_anim_state = LION_STOP;
