+++ src/game/objects/traps/pendulum.c
@@ -2,7 +2,7 @@
 
 #include "game/collide.h"
 #include "game/control.h"
-#include "game/objects/effects/blood.h"
+#include "game/effects/blood.h"
 #include "game/random.h"
 #include "global/vars.h"
 
@@ -38,7 +38,7 @@ void Pendulum_Control(int16_t item_num)
         int32_t z = g_LaraItem->pos.z + (Random_GetControl() - 0x4000) / 256;
         int32_t y = g_LaraItem->pos.y - Random_GetControl() / 44;
         int32_t d = g_LaraItem->pos.y_rot + (Random_GetControl() - 0x4000) / 8;
-        Blood_Spawn(x, y, z, g_LaraItem->speed, d, g_LaraItem->room_number);
+        Effect_Blood(x, y, z, g_LaraItem->speed, d, g_LaraItem->room_number);
     }
 
     FLOOR_INFO *floor =
